@@ -39,8 +39,8 @@
 	"dependencies": {
 		"marked": "~0.3.6",
 		"uglify-es": "~3.0.28",
-		"somes": "~0.10.112",
-		"typescript": "~3.7.4"
+		"suark": "~0.1.0",
+		"typescript": "~5.4.5"
 	},
 	"devDependencies": {
 		"@types/node": "~14.14.22"
